fix no folder list issue
